moved running of simulation to library
